cmd/tailscale/jsonoutput: hoist jsonoutput out of cli package
Flatten the cmd/tailscale package hierarchy by extracting the
jsonoutput package out of the cmd/tailscale/cli package.

Updates #cleanup
