+++ cmd/tailscale/jsonoutput/dns.go
