+++ cmd/tailscale/jsonoutput/example_schemaversion_test.go
@@ -7,7 +7,7 @@ import (
 	"flag"
 	"fmt"
 
-	"tailscale.com/cmd/tailscale/cli/jsonoutput"
+	"tailscale.com/cmd/tailscale/jsonoutput"
 )
 
 var args struct {
