+++ cmd/tailscale/jsonoutput/format.go
