+++ cmd/tailscale/jsonoutput/format_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"tailscale.com/cmd/tailscale/cli/jsonoutput"
+	"tailscale.com/cmd/tailscale/jsonoutput"
 )
 
 func TestFormat(t *testing.T) {
