+++ cmd/tailscale/jsonoutput/jsonoutput.go
