+++ cmd/tailscale/jsonoutput/tailnet-lock-log.go
