+++ cmd/tailscale/jsonoutput/tailnet-lock-status.go
